+++ packages/core/handlers/auth.js
@@ -27,7 +27,7 @@ async function onOAuthCallback({ platform, hostname, tokenUrl, callbackUri, apiU
     const authHeader = `Bearer ${accessToken}`;
     const { successful, platformUserInfo, returnMessage } = await platformModule.getUserInfo({ authHeader, tokenUrl, apiUrl, hostname, username, callbackUri, query, data });
     if (successful) {
-        const userInfo = await saveUserInfo({
+        let userInfo = await saveUserInfo({
             platformUserInfo,
             platform,
             tokenUrl,
@@ -36,8 +36,12 @@ async function onOAuthCallback({ platform, hostname, tokenUrl, callbackUri, apiU
             hostname: platformUserInfo?.overridingHostname ? platformUserInfo.overridingHostname : hostname,
             accessToken,
             refreshToken,
-            tokenExpiry: expires
+            tokenExpiry: expires,
+            rcAccountId: query.rcAccountId
         });
+        if (platformModule.postSaveUserInfo) {
+            userInfo = await platformModule.postSaveUserInfo({ userInfo, oauthApp });
+        }
         return {
             userInfo,
             returnMessage
@@ -56,12 +60,15 @@ async function onApiKeyLogin({ platform, hostname, apiKey, additionalInfo }) {
     const basicAuth = platformModule.getBasicAuth({ apiKey });
     const { successful, platformUserInfo, returnMessage } = await platformModule.getUserInfo({ authHeader: `Basic ${basicAuth}`, hostname, additionalInfo, apiKey });
     if (successful) {
-        const userInfo = await saveUserInfo({
+        let userInfo = await saveUserInfo({
             platformUserInfo,
             platform,
             hostname,
             accessToken: platformUserInfo.overridingApiKey ?? apiKey
         });
+        if (platformModule.postSaveUserInfo) {
+            userInfo = await platformModule.postSaveUserInfo({ userInfo });
+        }
         return {
             userInfo,
             returnMessage
@@ -75,7 +82,7 @@ async function onApiKeyLogin({ platform, hostname, apiKey, additionalInfo }) {
     }
 }
 
-async function saveUserInfo({ platformUserInfo, platform, hostname, accessToken, refreshToken, tokenExpiry }) {
+async function saveUserInfo({ platformUserInfo, platform, hostname, accessToken, refreshToken, tokenExpiry, rcAccountId }) {
     const id = platformUserInfo.id;
     const name = platformUserInfo.name;
     const existingUser = await UserModel.findByPk(id);
@@ -91,6 +98,7 @@ async function saveUserInfo({ platformUserInfo, platform, hostname, accessToken,
                 accessToken,
                 refreshToken,
                 tokenExpiry,
+                rcAccountId,
                 platformAdditionalInfo: {
                     ...existingUser.platformAdditionalInfo, // keep existing platformAdditionalInfo
                     ...platformAdditionalInfo,
@@ -113,6 +121,7 @@ async function saveUserInfo({ platformUserInfo, platform, hostname, accessToken,
                     accessToken,
                     refreshToken,
                     tokenExpiry,
+                    rcAccountId,
                     platformAdditionalInfo,
                     userSettings: userWithOldID.userSettings
                 });
@@ -128,6 +137,7 @@ async function saveUserInfo({ platformUserInfo, platform, hostname, accessToken,
                     accessToken,
                     refreshToken,
                     tokenExpiry,
+                    rcAccountId,
                     platformAdditionalInfo,
                     userSettings: {}
                 });
@@ -143,6 +153,7 @@ async function saveUserInfo({ platformUserInfo, platform, hostname, accessToken,
                 accessToken,
                 refreshToken,
                 tokenExpiry,
+                rcAccountId,
                 platformAdditionalInfo,
                 userSettings: {}
             });