+++ packages/core/handlers/log.js
@@ -7,8 +7,10 @@ const errorMessage = require('../lib/generalErrorMessage');
 const { composeCallLog, getLogFormatType } = require('../lib/callLogComposer');
 const adapterRegistry = require('../adapter/registry');
 const { LOG_DETAILS_FORMAT_TYPE } = require('../lib/constants');
+const { NoteCache } = require('../models/dynamo/noteCacheSchema');
+const moment = require('moment');
 
-async function createCallLog({ platform, userId, incomingData }) {
+async function createCallLog({ platform, userId, incomingData, hashedAccountId, isFromSSCL }) {
     try {
         const existingCallLog = await CallLogModel.findOne({
             where: {
@@ -39,7 +41,13 @@ async function createCallLog({ platform, userId, incomingData }) {
         const platformModule = adapterRegistry.getAdapter(platform);
         const callLog = incomingData.logInfo;
         const additionalSubmission = incomingData.additionalSubmission;
-        const note = incomingData.note;
+        let note = incomingData.note;
+        if (isFromSSCL) {
+            const noteCache = await NoteCache.get({ sessionId: incomingData.logInfo.sessionId });
+            if (noteCache) {
+                note = noteCache.note;
+            }
+        }
         const aiNote = incomingData.aiNote;
         const transcript = incomingData.transcript;
         const authType = platformModule.getAuthType();
@@ -104,7 +112,9 @@ async function createCallLog({ platform, userId, incomingData }) {
             additionalSubmission,
             aiNote,
             transcript,
-            composedLogDetails
+            composedLogDetails,
+            hashedAccountId,
+            isFromSSCL
         });
         if (logId) {
             await CallLogModel.create({
@@ -279,7 +289,7 @@ async function getCallLog({ userId, sessionIds, platform, requireDetails }) {
     }
 }
 
-async function updateCallLog({ platform, userId, incomingData }) {
+async function updateCallLog({ platform, userId, incomingData, hashedAccountId, isFromSSCL }) {
     try {
         const existingCallLog = await CallLogModel.findOne({
             where: {
@@ -335,7 +345,11 @@ async function updateCallLog({ platform, userId, incomingData }) {
                         sessionId: existingCallLog.sessionId,
                         startTime: incomingData.startTime,
                         duration: incomingData.duration,
-                        result: incomingData.result
+                        result: incomingData.result,
+                        direction: incomingData.direction,
+                        from: incomingData.from,
+                        to: incomingData.to,
+                        legs: incomingData.legs || [],
                     },
                     contactInfo: null, // Not needed for updates
                     user,
@@ -363,9 +377,12 @@ async function updateCallLog({ platform, userId, incomingData }) {
                 result: incomingData.result,
                 aiNote: incomingData.aiNote,
                 transcript: incomingData.transcript,
+                legs: incomingData.legs || [],
                 additionalSubmission: incomingData.additionalSubmission,
                 composedLogDetails,
-                existingCallLogDetails  // Pass the fetched details to avoid duplicate API calls
+                existingCallLogDetails,  // Pass the fetched details to avoid duplicate API calls
+                hashedAccountId,
+                isFromSSCL
             });
             return { successful: true, logId: existingCallLog.thirdPartyLogId, updatedNote, returnMessage, extraDataTracking };
         }
@@ -581,7 +598,19 @@ async function createMessageLog({ platform, userId, incomingData }) {
     }
 }
 
+async function saveNoteCache({ sessionId, note }) {
+    try {
+        const now = moment();
+        const noteCache = await NoteCache.create({ sessionId, note, ttl: now.unix() + 3600 });
+        return { successful: true, returnMessage: 'Note cache saved' };
+    } catch (e) {
+        console.error(`Error saving note cache: ${e.stack}`);
+        return { successful: false, returnMessage: 'Error saving note cache' };
+    }
+}
+
 exports.createCallLog = createCallLog;
 exports.updateCallLog = updateCallLog;
 exports.createMessageLog = createMessageLog;
 exports.getCallLog = getCallLog;
+exports.saveNoteCache = saveNoteCache;