+++ docs/release-notes.md
@@ -1,5 +1,32 @@
 # Release notes
 
+## 1.6.5:
+
+- New: Support call journey in call logging details with server side logging
+- Better: Tabs orders updated
+- Fix: Date/Time display issue
+
+## 1.6.4:
+- New Add RingCentral user name and phone number in call details settings
+- Better: Server-side call logging can now upload agent notes from the Chrome extension immediately after the call
+- Better: More intuitive date time user setting format
+- Fix: Return error message if server-side call logging is already created
+
+## 1.6.3:
+
+- New: Date time format in call details settings
+- Better: When doing server-side call logging, agent notes input from Chrome extension won't be overwritten by system
+- Better: When doing server-side call logging, agent notes input from Chrome extension will be auto uploaded
+- Fix: User report tab can now be toggled ON/OFF
+- Fix: Time format now follows operating system setting 
+- Bullhorn: Server-side call logging will only use API user session if it's configured with username and password
+
+## 1.6.2:
+
+- Fix: User session auth button now shows correct state
+- Fix: Auto log is not disabled by admin-only server side logging anymore
+- Fix: Playload too large issue
+
 ## 1.6.1:
 
 - New: Support to logging AI notes for no admin users in server side logging